implemented the first version of merging logic.
When a router receives a packet, it updates the latest received packet index for each recipient. It then checks whether there is merging opportunities such that there are other streams with lower latest received packet index.  If so, the router merges there two streams by sending one packet for all the merged streams.

In addition if there are multiple streams to merge, we always merge with the "top" stream (i.e., the stream that arrives at the router the earliest). This is a tradeoff since if we want to save the most packets, we need to record each merging point of other streams, which might kill the performance on the router.

Note that in this case some merged recipients will receive packets out of order (i.e.,  packets sent later by the server arrive earlier since they are merged by the router).  We need to have a way to buffer these out of order packets and send them to the player otherwise the player will fault out. We implemented a buffer at the player to address this problem. This implies that we need to change player applications, which is not desirable.  A solution is to implement the buffer at the edge routers subject to resource limitations.  We need to address this problem later.